feat: implement application lifecycle and window management
