+++ apps/desktop/layer/main/src/lib/tray.ts
@@ -4,8 +4,8 @@ import { app, Menu, nativeImage, Tray } from "electron"
 import { isMacOS, isMAS, isWindows } from "~/env"
 import { getTrayIconPath } from "~/helper"
 import { logger, revealLogFile } from "~/logger"
+import { WindowManager } from "~/manager/window"
 import { checkForAppUpdates } from "~/updater"
-import { WindowManager } from "~/window"
 
 import { getDockCount } from "./dock"
 import { t } from "./i18n"
