+++ apps/desktop/layer/main/src/manager/window.ts
@@ -9,15 +9,15 @@ import type { BrowserWindowConstructorOptions } from "electron"
 import { app, BrowserWindow, screen, shell } from "electron"
 import type { Event } from "electron/main"
 
-import { START_IN_TRAY_ARGS } from "./constants/app"
-import { isMacOS, isWindows, isWindows11 } from "./env"
-import { filePathToAppUrl, getIconPath } from "./helper"
-import { t } from "./lib/i18n"
-import { store } from "./lib/store"
-import { getTrayConfig } from "./lib/tray"
-import { refreshBound } from "./lib/utils"
-import { logger } from "./logger"
-import { loadDynamicRenderEntry } from "./updater/hot-updater"
+import { START_IN_TRAY_ARGS } from "~/constants/app"
+import { isMacOS, isWindows, isWindows11 } from "~/env"
+import { filePathToAppUrl, getIconPath } from "~/helper"
+import { t } from "~/lib/i18n"
+import { store } from "~/lib/store"
+import { getTrayConfig } from "~/lib/tray"
+import { refreshBound } from "~/lib/utils"
+import { logger } from "~/logger"
+import { loadDynamicRenderEntry } from "~/updater/hot-updater"
 
 const __dirname = fileURLToPath(new URL(".", import.meta.url))
