--- apps/desktop/layer/main/src/renderer-handlers.ts
@@ -1,5 +1,0 @@
-export type RendererHandlers = {
-  invalidateQuery: (key: (string | number | undefined)[]) => void
-  updateDownloaded: () => void
-  navigateEntry: (options: { feedId: string; entryId: string; view: number }) => void
-}
